frontend/mock_api/pyproject.toml
[tool.poetry]
name = "mock-api"
version = "0.1.0"
description = "Mock API for Median Code with Clerk authentication"
authors = ["Median Code Team"]
readme = "README.md"
packages = [{include = "mock_api", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.3"
httpx = "^0.26.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
cryptography = "^42.0.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.3"

# Note: Use 'poetry run pytest' or 'poetry run uvicorn mock_api.main:app --reload' instead of scripts

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]
